feat: rename Arch to ArchType
