+++ lib/data/model/swagger/model/base_swagger_model_response.dart
@@ -3,7 +3,7 @@ import 'package:onix_flutter_bricks/data/model/swagger/model/swagger_model_respo
 import 'package:onix_flutter_bricks/data/model/swagger/model/swagger_model_response_v2.dart';
 import 'package:onix_flutter_bricks/data/model/swagger/model/swagger_model_response_v3.dart';
 import 'package:onix_flutter_bricks/data/model/swagger/model_variable/base_swagger_model_variable_response.dart';
-import 'package:onix_flutter_bricks/domain/entity/arch/arch_type.dart';
+import 'package:onix_flutter_bricks/domain/entity/arch_type/arch_type.dart';
 
 abstract class BaseSwaggerModelResponse {
   final String name;
