+++ lib/data/model/swagger/path/base_swagger_path_response.dart
@@ -8,7 +8,7 @@ import 'package:onix_flutter_bricks/data/model/swagger/path/swagger_path_respons
 import 'package:onix_flutter_bricks/data/model/swagger/path/swagger_path_response_v3.dart';
 import 'package:onix_flutter_bricks/data/model/swagger/types/swagger_request_type.dart';
 import 'package:onix_flutter_bricks/data/model/swagger/types/swagger_response_type.dart';
-import 'package:onix_flutter_bricks/domain/entity/arch/arch_type.dart';
+import 'package:onix_flutter_bricks/domain/entity/arch_type/arch_type.dart';
 
 abstract class BaseSwaggerPathResponse {
   final String path;
