+++ lib/data/source/remote/swagger/swagger_remote_source.dart
@@ -1,5 +1,5 @@
 import 'package:onix_flutter_bricks/data/model/swagger/swagger_response.dart';
-import 'package:onix_flutter_bricks/domain/entity/arch/arch_type.dart';
+import 'package:onix_flutter_bricks/domain/entity/arch_type/arch_type.dart';
 
 // ignore: one_member_abstracts
 abstract interface class SwaggerRemoteSource {
