+++ lib/domain/entity/arch_type/arch_type.dart
