+++ lib/domain/repository/swagger_repository.dart
@@ -1,4 +1,4 @@
-import 'package:onix_flutter_bricks/domain/entity/arch/arch_type.dart';
+import 'package:onix_flutter_bricks/domain/entity/arch_type/arch_type.dart';
 import 'package:onix_flutter_bricks/domain/entity/component/components.dart';
 import 'package:onix_flutter_core/onix_flutter_core.dart';
