+++ lib/domain/service/file_generator_service/screen_generators/gen/mixins/di_content_mixin.dart
@@ -1,6 +1,6 @@
 import 'dart:io';
 
-import 'package:onix_flutter_bricks/domain/entity/arch/arch_type.dart';
+import 'package:onix_flutter_bricks/domain/entity/arch_type/arch_type.dart';
 import 'package:onix_flutter_bricks/domain/service/base/base_generation_service.dart';
 import 'package:onix_flutter_bricks/domain/service/file_generator_service/screen_generators/params/screen_generator_params.dart';
 import 'package:recase/recase.dart';
